Metodos rest

+++ src/main/java/com/unbosque/edu/co/repository/AuditoriaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.unbosque.edu.co.entity.Auditory;
 
-public interface AuditoriaRepository extends JpaRepository<Auditory, Long> {
+public interface AuditoriaRepository extends JpaRepository<Auditory, Integer> {
     // Puedes agregar métodos específicos de consulta si los necesitas
 }
