+++ src/main/java/com/unbosque/edu/co/repository/ClienteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.unbosque.edu.co.entity.Client;
 
-public interface ClienteRepository extends JpaRepository<Client, Long> {
+public interface ClienteRepository extends JpaRepository<Client, Integer> {
     // Puedes agregar métodos específicos de consulta si los necesitas
 }
