+++ src/main/java/com/unbosque/edu/co/repository/VehiculoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.unbosque.edu.co.entity.Vehicle;
 
-public interface VehiculoRepository extends JpaRepository<Vehicle, Long> {
+public interface VehiculoRepository extends JpaRepository<Vehicle, Integer> {
     // Puedes agregar métodos específicos de consulta si los necesitas
 }
